backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Anaskhan", email = "anaskhanx101@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.116.2",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]

[project.scripts]
backend = "backend:main"

[build-system]
requires = ["uv_build>=0.8.3,<0.9.0"]
build-backend = "uv_build"
